Final
Final
Unfortuantely, I failed to implement a database to save the scores in this app. I was having trouble applying it to a strict list of 9 objects, so I went back to my list of 9 integers instead, because I would rather submit something functional than something that is not functional.